+++ src/main/java/com/gamemoonchul/application/PostService.java
@@ -27,6 +27,7 @@ public class PostService {
     private final MatchUserService matchUserService;
     private final VoteOptionRepository voteOptionRepository;
 
+
     public PostResponseDto upload(PostUploadRequest request, Member member) {
         Post entity = PostConverter.requestToEntity(request, member);
         Post savedPost = postRepository.save(entity);
@@ -50,23 +51,5 @@ public class PostService {
                 .toList();
         List<VoteOptions> savedVoteOptions = voteOptionRepository.saveAll(voteOptions);
         post.addVoteOptions(voteOptions);
-    }
-
-    @Transactional
-    public String delete(Long postId, Member member) {
-        Post post = postRepository.findById(postId)
-                .orElseThrow(() -> {
-                            log.error(PostStatus.POST_NOT_FOUND.getMessage());
-                            return new NotFoundException(PostStatus.POST_NOT_FOUND);
-                        }
-                );
-        if (member.getId()
-                .equals(post.getMember()
-                        .getId())) {
-            postRepository.delete(post);
-            return "Delete Complete";
-        }
-        log.error(PostStatus.UNAUTHORIZED_REQUEST.getMessage());
-        throw new UnauthorizedException(PostStatus.UNAUTHORIZED_REQUEST);
     }
 }
